# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "77Sol - gerenciamento de projetos",
        "description": "Está documentação fornece os endpoints e detalhes para a integração com o sistema de gertão de projetos da 77Sol.",
        "version": "1.0"
    },
    "paths": {
...
        "/api/customers/{id}": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Retorna um cliente cadastrado",
                "description": "Retorna um cliente cadastrado",
                "operationId": "472d307317b0276e26ae47975377a4cd",
                "responses": {
                    "200": {
                        "description": "Retorna um cliente pelo ID"
                    }
                }
            },
            "put": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Editar cliente",
                "description": "Editar Cliente",
                "operationId": "88724fadb81751fd3cf31633c61bfab3",
                "requestBody": {
                    "content": {
                        "multipart/formdata": {
                            "schema": {
                                "required": [
                                    "name",
                                    "document"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "text"
                                    },
                                    "emails": {
                                        "type": "text"
                                    },
                                    "contact_numbers": {
                                        "type": "text"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Editar um cliente cadastrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Deleta um cliente cadastrado",
                "description": "Deleta um cliente cadastrado",
                "operationId": "1b15a111914cf397567ffbdd2d91abaf",
                "responses": {
                    "200": {
                        "description": "Deleta um cliente pelo ID"
                    }
                }
            }
        },
...
    }
}